Added a simple editor with syntax highlight of Objective-C. It's still needed to create a dialog to allow customization of the colours applied; right now it relies on constants.
Also added a bunch of classes to read Xcode 3.x project files. The original intention was to use these to store most of the project's configuration, allowing easy exchange of Xcode projects with Eclipse. But it can also be used to import Xcode projects if the original idea were not good.

git-svn-id: https://objectiveclipse.googlecode.com/svn/trunk@25 b0dc983c-f627-11dd-ada4-5153b8187bf2